requirements.txt
fastapi==0.88.0
requests==2.27.1
Pillow==9.0.0
pyyaml==5.3.1
cryptography==36.0.0
psycopg2-binary==2.9.5
uvicorn==0.20.0
python-multipart==0.0.5
